Update weight
fixed a bug where character in sequence was not a match for one letter abbreviation in AA_table. Before the weight was returned as NA. However, a warning should be made to let the user know that there are some characters in the amino acid sequence that are not mapped to a weight value. Example: X is not a valid one letter AA abbreviation and thus it's weight is NA which when adding NA to a total number the result is NA. The change in the code filters out all NA's in a character sequence.